+++ src/sdk/data/classes/connext-token.ts
@@ -1,0 +1,8 @@
+export class ConnextToken {
+    symbol: string;
+    address: string;
+    decimals: number;
+    chainId: number;
+    name: string;
+    icon: string;
+}