+++ src/sdk/data/classes/connext-transaction-status.ts
@@ -1,0 +1,6 @@
+export class ConnextTransactionStatus {
+    status: string;
+    transactionHash?: string;
+    connextscanUrl: string;
+    transferId?: string;
+}