+++ src/sdk/dto/get-connext-supported-assets.dto.ts
@@ -1,0 +1,8 @@
+import { IsInt, IsOptional, IsPositive } from 'class-validator';
+
+export class GetConnextSupportedAssetsDto {
+  @IsOptional()
+  @IsPositive()
+  @IsInt()
+  chainId?: number = null;
+}